BZ 805224 limit action to be executable only every 2 minutes (cherry picked from commit b3f351d5610050e8291c42c65d6b4506342d8f47)

+++ src/app/models/instance.rb
@@ -478,7 +478,12 @@ class Instance < ActiveRecord::Base
   end
 
   def reboot(user)
-    do_operation(user, 'reboot')
+    if tasks.where("action = :action AND time_submitted > :time_ago",
+       {:action => "reboot", :time_ago => 2.minutes.ago}).present?
+      raise I18n.t("instances.errors.reboot_already_scheduled")
+    else
+      do_operation(user, 'reboot')
+    end
   end
 
   def forced_stop(user)
